refactor:  centralized power states
