+++ pkg/wsman/cim/associatedpower/constants.go
@@ -1,0 +1,10 @@
+/*********************************************************************
+ * Copyright (c) Intel Corporation 2026
+ * SPDX-License-Identifier: Apache-2.0
+ **********************************************************************/
+
+package associatedpower
+
+const (
+	CIMAssociatedPowerManagementService string = "CIM_AssociatedPowerManagementService"
+)
