+++ pkg/wsman/cim/associatedpower/types.go
@@ -9,6 +9,7 @@ import (
 	"encoding/xml"
 
 	"github.com/device-management-toolkit/go-wsman-messages/v2/internal/message"
+	"github.com/device-management-toolkit/go-wsman-messages/v2/pkg/wsman/cim/models"
 	"github.com/device-management-toolkit/go-wsman-messages/v2/pkg/wsman/client"
 	"github.com/device-management-toolkit/go-wsman-messages/v2/pkg/wsman/common"
 )
@@ -32,15 +33,15 @@ type (
 		AssociatedPowerManagementServiceItems []CIM_AssociatedPowerManagementService `xml:"Items>CIM_AssociatedPowerManagementService"`
 	}
 	CIM_AssociatedPowerManagementService struct {
-		AvailableRequestedPowerStates []AvailableRequestedPowerStates `xml:"AvailableRequestedPowerStates,omitempty"` // AvailableRequestedPowerStates indicates the possible values for the PowerState parameter of the method RequestPowerStateChange, used to initiate a power state change.
-		PowerState                    PowerState                      `xml:"PowerState,omitempty"`                    // The current power state of the associated Managed System Element.
-		OtherPowerState               string                          `xml:"OtherPowerState,omitempty"`               // A string describing the additional power management state of the element, used when the PowerState is set to the value 1, "Other".
-		RequestedPowerState           RequestedPowerState             `xml:"RequestedPowerState,omitempty"`           // The desired or the last requested power state of the associated Managed System Element.
-		OtherRequestedPowerState      string                          `xml:"OtherRequestedPowerState,omitempty"`      // A string describing the additional power management state of the element, used when the RequestedPowerState is set to the value 1, "Other".
-		PowerOnTime                   string                          `xml:"PowerOnTime,omitempty"`                   // The time when the element will be powered on again, used when the RequestedPowerState has the value 2, "On", 5, "Power Cycle (Off - Soft)" or 6, "Power Cycle (Off - Hard)".
-		TransitioningToPowerState     TransitioningToPowerState       `xml:"TransitioningToPowerState,omitempty"`     // TransitioningToPowerState indicates the target power state to which the system is transitioning.
-		ServiceProvided               ServiceProvided                 // The Service that is available.
-		UserOfService                 UserOfService                   // The ManagedElement that can use the Service.
+		AvailableRequestedPowerStates []models.AvailableRequestedPowerStates `xml:"AvailableRequestedPowerStates,omitempty"` // AvailableRequestedPowerStates indicates the possible values for the PowerState parameter of the method RequestPowerStateChange, used to initiate a power state change.
+		PowerState                    models.PowerState                      `xml:"PowerState,omitempty"`                    // The current power state of the associated Managed System Element.
+		OtherPowerState               string                                 `xml:"OtherPowerState,omitempty"`               // A string describing the additional power management state of the element, used when the PowerState is set to the value 1, "Other".
+		RequestedPowerState           models.RequestedPowerState             `xml:"RequestedPowerState,omitempty"`           // The desired or the last requested power state of the associated Managed System Element.
+		OtherRequestedPowerState      string                                 `xml:"OtherRequestedPowerState,omitempty"`      // A string describing the additional power management state of the element, used when the RequestedPowerState is set to the value 1, "Other".
+		PowerOnTime                   string                                 `xml:"PowerOnTime,omitempty"`                   // The time when the element will be powered on again, used when the RequestedPowerState has the value 2, "On", 5, "Power Cycle (Off - Soft)" or 6, "Power Cycle (Off - Hard)".
+		TransitioningToPowerState     models.TransitioningToPowerState       `xml:"TransitioningToPowerState,omitempty"`     // TransitioningToPowerState indicates the target power state to which the system is transitioning.
+		ServiceProvided               ServiceProvided                        // The Service that is available.
+		UserOfService                 UserOfService                          // The ManagedElement that can use the Service.
 	}
 	ServiceProvided struct {
 		XMLName             xml.Name `xml:"ServiceProvided,omitempty"`
@@ -61,13 +62,4 @@ type (
 		ResourceURI string   `xml:"ResourceURI,omitempty"`
 		SelectorSet SelectorSet
 	}
-
-	// AvailableRequestedPowerStates indicates the possible values for the PowerState parameter of the method RequestPowerStateChange, used to initiate a power state change.
-	AvailableRequestedPowerStates int
-	// PowerState indicates the current power state of the associated Managed System Element.
-	PowerState int
-	// RequestedPowerState indicates the desired or the last requested power state of the associated Managed System Element.
-	RequestedPowerState int
-	// TransitioningToPowerState indicates the target power state to which the system is transitioning.
-	TransitioningToPowerState int
 )
