+++ pkg/wsman/cim/models/powerstates_test.go
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  **********************************************************************/
 
-package associatedpower
+package models
 
 import "testing"
