+++ pkg/wsman/cim/service/availabletoelement_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/device-management-toolkit/go-wsman-messages/v2/internal/message"
+	"github.com/device-management-toolkit/go-wsman-messages/v2/pkg/wsman/cim/models"
 	"github.com/device-management-toolkit/go-wsman-messages/v2/pkg/wsman/common"
 	"github.com/device-management-toolkit/go-wsman-messages/v2/pkg/wsman/wsmantesting"
 )
@@ -91,7 +92,7 @@ func TestNegativeAvailableToElement(t *testing.T) {
 						XMLName: xml.Name{Space: "http://schemas.xmlsoap.org/ws/2004/09/enumeration", Local: "PullResponse"},
 						AssociatedPowerManagementService: []CIM_AssociatedPowerManagementService{
 							{
-								AvailableRequestedPowerStates: []AvailableRequestedPowerStates{10},
+								AvailableRequestedPowerStates: []models.AvailableRequestedPowerStates{10},
 								PowerState:                    2,
 								ServiceProvided: ServiceProvided{
 									XMLName: xml.Name{Space: "http://schemas.xmlsoap.org/ws/2004/08/addressing", Local: "ServiceProvided"},
