+++ pkg/wsman/cim/service/constants.go
@@ -1,0 +1,10 @@
+/*********************************************************************
+ * Copyright (c) Intel Corporation 2023
+ * SPDX-License-Identifier: Apache-2.0
+ **********************************************************************/
+
+package service
+
+const (
+	CIMServiceAvailableToElement string = "CIM_ServiceAvailableToElement"
+)
