+++ pkg/wsman/cim/service/types.go
@@ -9,6 +9,7 @@ import (
 	"encoding/xml"
 
 	"github.com/device-management-toolkit/go-wsman-messages/v2/internal/message"
+	"github.com/device-management-toolkit/go-wsman-messages/v2/pkg/wsman/cim/models"
 	"github.com/device-management-toolkit/go-wsman-messages/v2/pkg/wsman/client"
 	"github.com/device-management-toolkit/go-wsman-messages/v2/pkg/wsman/common"
 )
@@ -32,10 +33,10 @@ type (
 		AssociatedPowerManagementService []CIM_AssociatedPowerManagementService `xml:"Items>CIM_AssociatedPowerManagementService"`
 	}
 	CIM_AssociatedPowerManagementService struct {
-		AvailableRequestedPowerStates []AvailableRequestedPowerStates `xml:"AvailableRequestedPowerStates,omitempty"` // AvailableRequestedPowerStates indicates the possible values for the PowerState parameter of the method RequestPowerStateChange, used to initiate a power state change.The values listed shall be a subset of the values contained in the RequestedPowerStatesSupported property of the CIM_PowerManagementCapabilities where the values selected are a function of the current power state of the system. This property shall be non-null if an implementation supports the advertisement of the set of possible values as a function of the current state. This property shall be null if an implementation does not support the advertisement of the set of possible values as a function of the current state.
-		PowerState                    PowerState                      `xml:"PowerState,omitempty"`                    // The current power state of the associated Managed System Element.
-		ServiceProvided               ServiceProvided                 // The Service that is available.
-		UserOfService                 UserOfService                   // The ManagedElement that can use the Service.
+		AvailableRequestedPowerStates []models.AvailableRequestedPowerStates `xml:"AvailableRequestedPowerStates,omitempty"` // AvailableRequestedPowerStates indicates the possible values for the PowerState parameter of the method RequestPowerStateChange, used to initiate a power state change.The values listed shall be a subset of the values contained in the RequestedPowerStatesSupported property of the CIM_PowerManagementCapabilities where the values selected are a function of the current power state of the system. This property shall be non-null if an implementation supports the advertisement of the set of possible values as a function of the current state. This property shall be null if an implementation does not support the advertisement of the set of possible values as a function of the current state.
+		PowerState                    models.PowerState                      `xml:"PowerState,omitempty"`                    // The current power state of the associated Managed System Element.
+		ServiceProvided               ServiceProvided                        // The Service that is available.
+		UserOfService                 UserOfService                          // The ManagedElement that can use the Service.
 	}
 	ServiceProvided struct {
 		XMLName             xml.Name `xml:"ServiceProvided,omitempty"`
@@ -52,9 +53,4 @@ type (
 		ResourceURI string   `xml:"ResourceURI,omitempty"`
 		SelectorSet message.SelectorSet
 	}
-
-	// AvailableRequestedPowerStates indicates the possible values for the PowerState parameter of the method RequestPowerStateChange, used to initiate a power state change.
-	AvailableRequestedPowerStates int
-	// PowerState indicates the current power state of the associated Managed System Element.
-	PowerState int
 )
